fix: rename evaluate seniority hook
